+++ src/TypeProviderInstantiation.Experimental.fs
@@ -3,34 +3,30 @@
 open ProviderImplementation.ProvidedTypes
 
 type ApiaryProviderArgs = 
-    { ApiName : string 
-      SpecialNames : string}
+    { ApiName : string }
 
 type TypeProviderInstantiation = 
     | Apiary of ApiaryProviderArgs
 
-    member x.GenerateType resolutionFolder runtimeAssembly platform =
+    member x.GenerateType resolutionFolder runtimeAssembly =
         let f, args =
             match x with
             | Apiary x ->
                 (fun cfg -> new ApiaryProvider(cfg) :> TypeProviderForNamespaces),
-                [| box x.ApiName
-                   box x.SpecialNames |] 
-        Debug.generate resolutionFolder runtimeAssembly platform f args
+                [| box x.ApiName |] 
+        Debug.generate resolutionFolder runtimeAssembly f args
 
     override x.ToString() =
         match x with
         | Apiary x -> ["Apiary"
-                       x.ApiName
-                       x.SpecialNames]
+                       x.ApiName]
         |> String.concat ","
 
     static member Parse (line:string) =
         let args = line.Split [|','|]
         match args.[0] with
         | "Apiary" ->
-            Apiary { ApiName = args.[1]
-                     SpecialNames = args.[2].Replace(';', ',') }
+            Apiary { ApiName = args.[1] }
         | _ -> failwithf "Unknown: %s" args.[0]
 
 open System.Runtime.CompilerServices