+++ RELEASE_NOTES.md
@@ -138,11 +138,17 @@
 #### 2.0.0-alpha6 - February 4 2014
 * JsonValue.Post() allows to post the JSON to the specified uri using HTTP.
 
-#### 2.0.0-alpha7 - Unreleased
+#### 2.0.0-alpha7 - February 20 2014
 * Improved name generation algorithm to cope better with acronymns.
 * Fixed wrong singularization of words ending with 'uses'.
 * Fixed handling of repeated one letter names.
 * Improve HTTP error messages.
 * Support for more api patterns in ApiaryProvider.
 * Tolerate invalid json and missing data in apiary specifications.
-* Added 'SpecialNames' parameter to ApiaryProvider to allow taking concrete examples in the apiary specification and turning them to api parameters.
+* Improved naming of generated types.
+* Fixed 'SampleIsList' to work with xml and json spanning multiple lines.
+* Fixed handling of nested arrays in JsonProvider.
+* Fixed handling of optional arrays in JsonProvider.
+
+#### 2.0.0-alpha8 - Unreleased
+* Mono fixes.